+++ components/Forms.php
@@ -20,6 +20,7 @@ use yii\base\Component;
 use wdmg\base\models\DynamicModel;
 use yii\base\InvalidArgumentException;
 use yii\helpers\ArrayHelper;
+use yii\helpers\Html;
 
 class Forms extends Component
 {
@@ -37,6 +38,11 @@ class Forms extends Component
         $this->model = new \wdmg\forms\models\Forms;
     }
 
+    /**
+     * @param null $widget
+     * @param null $id
+     * @return string|null
+     */
     public function build($widget = null, $id = null)
     {
         if (is_null($id))
@@ -56,17 +62,16 @@ class Forms extends Component
             $output = '';
             if ($fields = $form->getFormsFields(['locale' => $locale], true, false)->all()) {
 
-                // @TODO: Add default controller for form submits and storage data
-                // @TODO: Add custom options for forms
-
-                /*$widget->action = '';
-                $widget->method = '';
                 $widget->options = [
-                    'id' => $form->alias
-                ];*/
+                    'id' => 'form-' . $form->id,
+                    'name' => $form->alias
+                ];
 
                 $model = new DynamicModel();
                 foreach ($fields as $field) {
+
+                    $formName = $this->getFormName($form->alias);
+                    $model->setFormName($formName);
                     $model->defineAttribute($field->attribute);
                     $model->addRule($field->attribute, $field->getValidator());
                     $model->setAttributeLabel([$field->attribute => $field->label]);
@@ -94,6 +99,121 @@ class Forms extends Component
         return null;
     }
 
+    public function submit($id = null, $data = null) {
+
+        if (is_null($id) ||is_null($data))
+            return null;
+
+        if (is_string($id))
+            $form = $this->model->getPublished(['alias' => $id], true);
+        else
+            $form = $this->model->getPublished(['id' => $id], true);
+
+        if ($form) {
+
+            // Check of current form exist and collect data
+            $collect = null;
+            $formName = $this->getFormName($form->alias);
+            if (isset($data[$formName]))
+                $collect[$formName] = $data[$formName];
+            else
+                return null;
+
+            $errors = [];
+            if (is_countable($collect) && $fields = $form->getFormsFields(['source_id' => null], true, false)->all()) {
+
+                $fields_ids = [];
+                $model = new DynamicModel();
+                foreach ($fields as $field) {
+
+                    $formName = $this->getFormName($form->alias);
+                    $model->setFormName($formName);
+                    $model->defineAttribute($field->attribute);
+                    $model->addRule($field->attribute, $field->getValidator());
+                    $model->setAttributeLabel([$field->attribute => $field->label]);
+
+                    $field_name = str_replace('-', '_', $field->name);
+                    $fields_ids[$field_name] = $field->id;
+                }
+
+                if ($model->load($collect) && $model->validate()) {
+
+                    if (!empty($model->errors))
+                        $errors[$formName] = $model->errors;
+
+                    $attributes = $model->getAttributes();
+
+
+                    if (is_countable($attributes)) {
+
+                        $hasError = false;
+
+                        $submit = new \wdmg\forms\models\Submits();
+                        $submit->form_id = $form->id;
+                        $submit->user_id = (Yii::$app->getUser()) ? Yii::$app->getUser()->id : null;
+                        $submit->access_token = Yii::$app->security->generateRandomString();
+                        $submit->status = 0;
+
+                        if ($submit->validate()) {
+
+                            if (!empty($submit->errors))
+                                $errors['Submits'] = $submit->errors;
+
+                            if ($submit->save()) {
+
+                                foreach ($attributes as $attribute => $value) {
+
+                                    $content = new \wdmg\forms\models\Content();
+                                    $content->input_id = ArrayHelper::getValue($fields_ids, $attribute);
+                                    $content->submit_id = $submit->id;
+                                    $content->value = $value;
+
+                                    if ($content->validate()) {
+                                        if (!$content->save()) {
+                                            $hasError = true;
+                                        }
+                                    } else {
+                                        if (!empty($content->errors)) {
+                                            $errors['Content'] = ArrayHelper::merge($content->errors, $errors);
+                                        }
+                                        $hasError = true;
+                                    }
+                                }
+
+                                if (!$hasError) {
+                                    $submit->status = 1;
+
+                                    if ($submit->update())
+                                        return true;
+
+                                }
+
+                                return (!empty($errors)) ? $errors : false;
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        return false;
+    }
+
+    private function getFormName($formAlias = null) {
+
+        if (is_null($formAlias))
+            return null;
+
+        return str_replace(' ', '', \yii\helpers\Inflector::camel2words($formAlias));
+    }
+
+    private function getFormAlias($formName = null) {
+
+        if (is_null($formName))
+            return null;
+
+        return str_replace(' ', '', \yii\helpers\Inflector::camel2id($formName));
+    }
 }
 
 ?>